requirements.txt
# Core Docling and PDF processing
docling>=2.0.0
pypdf2>=3.0.0

# OCR engines (local-first)
rapidocr-onnxruntime>=1.3.0
pytesseract>=0.3.10

# Embeddings and vector DB
sentence-transformers>=2.2.2
chromadb>=0.4.0

# CLIP for multimodal embeddings
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0

# LLM integration
openai>=1.0.0
ollama>=0.1.0

# Data processing and utilities
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Image processing
opencv-python>=4.8.0

# API and automation (optional for n8n integration)
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
ipython>=8.12.0
